matrMult: add optimal synchronized version
The synchronized approach to matrix multiplication was optimized.
This way the operation is as parallel as possible, if there are enough
threads available, every thread will do just one multiplication and add
its result to a cell in the output matrix.
Another version was added, that does not need synchronization, as every
cell of the output matrix is managed by a different thread. It will be
interesting to see the difference in performance of the two approaches